backend/requirements.txt
asgiref==3.4.1
cycler==0.10.0
Django==3.2.5
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filter==1.1.0
django-jsonfield==1.4.1
django-taggit==1.5.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
graphviz==0.17
kiwisolver==1.3.1
matplotlib==3.4.2
mysqlclient==2.0.3
numpy==1.21.1
Pillow==8.3.1
pydot3==1.0.9
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
